% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{ccAFv2_classifier}
\alias{ccAFv2_classifier}
\title{ccAFv2 classifier function}
\usage{
ccAFv2_classifier(norm_expVec)
}
\arguments{
\item{norm_expVec}{a double precision 1 x 861 vector of normalized gene expression values}
}
\value{
ccAFv2_classifier returns a 1 x 7 double precision vector of class probabilites
}
\description{
This function calls the C code C_ccAFv2.c to run the neural network classifier. 
ccAFv2_classifier returns a 1 x 7 double precision vector of predictions
}
