% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{SpatialDimPlot.ccAFv2}
\alias{SpatialDimPlot.ccAFv2}
\title{SpatialDimPlot of ccAFv2 predictions with standard colors}
\usage{
SpatialDimPlot.ccAFv2(seurat_obj, ...)
}
\arguments{
\item{Seurat}{object with ccAFv2 cell cycle states predicted.}
}
\value{
A DimPlot object that can be plotted.
}
\description{
This function plots the cell cycle onto a DimPlot for spatial data.
}
