% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{DimPlot.ccAFv2}
\alias{DimPlot.ccAFv2}
\title{DimPlot of ccAFv2 predictions with standard colors}
\usage{
DimPlot.ccAFv2(seurat_obj, ...)
}
\arguments{
\item{Seurat}{object that should have ccAFv2 cell cycle states predicted.}
}
\value{
A DimPlot object that can be plotted.
}
\description{
This function plots the cell cycle onto a DimPlot for single cell or nuclei data.
}
