% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{PrepareForCellCycleRegression}
\alias{PrepareForCellCycleRegression}
\title{Prepare expression module scores for regressing out the cell cycle}
\usage{
PrepareForCellCycleRegression(
  seurat_obj,
  assay = "SCT",
  species = "human",
  gene_id = "ensembl"
)
}
\arguments{
\item{seurat0:}{a seurat object must be supplied to classify, no default}

\item{assay:}{which seurat_obj assay to use for classification, helpful if data is prenormalized, default is 'SCT'}

\item{species:}{from which species did the samples originate, either 'human' or 'mouse', defaults to 'human'}

\item{gene_id:}{what type of gene ID is used, either 'ensembl' or 'symbol', defaults to 'ensembl'}

\item{spatial:}{whether the data is spatial, defaults to FALSE}
}
\value{
Seurat object with cell cycle state module scores appended to meta.data
}
\description{
This function computes moduel scores for each cell cycle state
}
