% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{ThresholdPlot}
\alias{ThresholdPlot}
\title{ThresholdPlot of ccAFv2 predictions with a range of thresholds using standard colors}
\usage{
ThresholdPlot(seurat_obj, ...)
}
\arguments{
\item{Seurat}{object with ccAFv2 cell cycle states predicited.}
}
\value{
A ggplot object that can be plotted.
}
\description{
This function plots the distribution of cell cycle predictions for a range of thresholds
as a barplot colorized using the standard cell cycle state colors.
}
